[step5] refactor: domain, controller,dto 패키지 생성, 관련 파일 위치 이동
