+++ src/main/kotlin/racingcar/domain/MoveStrategy.kt
@@ -1,4 +1,4 @@
-package racingcar.car.strategy
+package racingcar.domain
 
 interface MoveStrategy {
     fun isMovable(): Boolean
