+++ src/main/kotlin/racingcar/exception/GameOverException.kt
@@ -1,3 +1,3 @@
-package racingcar.car.exception
+package racingcar.exception
 
 class GameOverException : RuntimeException("게임을 더이상 할 수 없습니다")
