+++ src/main/kotlin/racingcar/exception/InvalidDriverNameException.kt
@@ -1,3 +1,3 @@
-package racingcar.car.exception
+package racingcar.exception
 
 class InvalidDriverNameException(name: String, limit: Int) : RuntimeException("이름은 ${limit}을 넘을 수 없습니다. `$name`")
