+++ src/test/kotlin/racingcar/car/CarTest.kt
@@ -2,6 +2,8 @@ package racingcar.car
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.domain.Car
+import racingcar.dto.MoveResults
 
 class CarTest {
     @Test
