+++ src/test/kotlin/racingcar/car/NameStrategyTest.kt
@@ -3,8 +3,8 @@ package racingcar.car
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import racingcar.car.exception.InvalidDriverNameException
-import racingcar.car.strategy.NameLengthLimitStrategy
+import racingcar.domain.NameLengthLimitStrategy
+import racingcar.exception.InvalidDriverNameException
 
 class NameStrategyTest {
     @Test
